requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --allow-unsafe --output-file=requirements.txt requirements.in
#
agentarchives==0.9.0
    # via -r requirements.in
amclient==1.3.0
    # via -r requirements.in
ammcpc==0.2.0
    # via -r requirements.in
asgiref==3.8.1
    # via django
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
bagit @ git+https://github.com/artefactual-labs/bagit-python.git@4b8fde73b4e631461bfd7add87e200500d40ca21
    # via -r requirements.in
brotli==1.1.0
    # via -r requirements.in
certifi==2024.2.2
    # via requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
clamd==1.0.2
    # via -r requirements.in
cryptography==42.0.7
    # via
    #   josepy
    #   mozilla-django-oidc
    #   pyopenssl
django==4.2.13
    # via
    #   -r requirements.in
    #   django-auth-ldap
    #   django-cas-ng
    #   django-csp
    #   mozilla-django-oidc
django-auth-ldap==4.8.0
    # via -r requirements.in
django-autoslug==1.9.9
    # via -r requirements.in
django-cas-ng==5.0.1
    # via -r requirements.in
django-csp==3.8
    # via -r requirements.in
django-forms-bootstrap==3.1.0
    # via -r requirements.in
django-prometheus==2.3.1
    # via -r requirements.in
django-shibboleth-remoteuser @ git+https://github.com/artefactual-labs/django-shibboleth-remoteuser.git@f08a7864d6130416c352981ccf318fff0fd5be58
    # via -r requirements.in
django-tastypie==0.14.7
    # via -r requirements.in
elasticsearch==6.8.2
    # via -r requirements.in
gearman3 @ git+https://github.com/artefactual-labs/python-gearman.git@b68efc868c7a494dd6a2d2e820fb098a6da9f797
    # via -r requirements.in
gevent==24.2.1
    # via -r requirements.in
greenlet==3.0.3
    # via gevent
grpcio==1.64.1
    # via -r requirements.in
gunicorn==22.0.0
    # via -r requirements.in
idna==3.7
    # via requests
importlib-metadata==7.1.0
    # via -r requirements.in
importlib-resources==6.4.0
    # via opf-fido
inotify-simple==1.3.5
    # via -r requirements.in
josepy==1.14.0
    # via mozilla-django-oidc
jsonschema==4.22.0
    # via -r requirements.in
jsonschema-specifications==2023.12.1
    # via jsonschema
lazy-paged-sequence==0.3
    # via -r requirements.in
lxml==5.2.2
    # via
    #   -r requirements.in
    #   ammcpc
    #   metsrw
    #   python-cas
metsrw==0.5.1
    # via -r requirements.in
mozilla-django-oidc==4.0.1
    # via -r requirements.in
mysqlclient==2.2.4
    # via agentarchives
olefile==0.47
    # via opf-fido
opf-fido @ git+https://github.com/artefactual-labs/fido.git@564ceb8018a8650fe931cf20e6780ee008e60fca
    # via -r requirements.in
orjson==3.10.1
    # via -r requirements.in
packaging==24.0
    # via gunicorn
prometheus-client==0.20.0
    # via
    #   -r requirements.in
    #   django-prometheus
protobuf==5.27.1
    # via -r requirements.in
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   python-ldap
pyasn1-modules==0.4.0
    # via python-ldap
pycparser==2.22
    # via cffi
pyopenssl==24.1.0
    # via josepy
python-cas==1.6.0
    # via django-cas-ng
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   django-tastypie
python-ldap==3.4.4
    # via
    #   -r requirements.in
    #   django-auth-ldap
python-mimeparse==1.6.0
    # via django-tastypie
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.1
    # via
    #   -r requirements.in
    #   agentarchives
    #   amclient
    #   mozilla-django-oidc
    #   opf-fido
    #   python-cas
rpds-py==0.18.1
    # via
    #   jsonschema
    #   referencing
six==1.16.0
    # via
    #   opf-fido
    #   python-cas
    #   python-dateutil
sqlparse==0.5.0
    # via django
typing-extensions==4.11.0
    # via asgiref
unidecode==1.3.8
    # via -r requirements.in
urllib3==2.2.1
    # via
    #   amclient
    #   elasticsearch
    #   requests
whitenoise==6.6.0
    # via -r requirements.in
zipp==3.18.2
    # via
    #   importlib-metadata
    #   importlib-resources
zope-event==5.0
    # via gevent
zope-interface==6.4
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
setuptools==69.5.1
    # via
    #   zope-event
    #   zope-interface
